Remove colored exception output
Trying to import IPython exception printer imports all kinds of things,
resulting in slow concert binary run-time without doing anything. If this is
really required it should be moved to a function that imports the formatter at
run-time. However, I strongly object this because the async module should be
free of any view related code.