Fix potential segfault
Passing the NumPy data pointer to the buffer does not work at least on one
machine. Since this always seemed to be to brittle, we added a new call to
ufo-core which allows data to be written to the host array. This means an
additional data copy but this is effectively the same as in the very first
version of the InjectProcess.