viewers: Use qsize() instead of empty() on queues
According to the multiprocessing module empty() is non-reliable, which means the
queue might become large and cause a lag in displaying images. It seems that
testing the queue size with qsize() works better.
However, the qsize() is also marked as non-reliable, so if the lag problem
appears again we need to implement something more sophisticated, e.g. a control
channel between the main and the display process. This will not be done now
becaue on the long-run the matplotlib-based viewers will be deprecated.