viewers: Do not use pause to process GUI events
Becuase it caused the window to stay on top. Also reimplement simple
updater's close behavior needed by this change. It could happen that we
are closed in the middle of the process_image method which would put the
drawing objects to None after they are checked for it and that causes no
good.